Removes real tests from non-widget files, adds some new found tests.
