+++ spyder/utils/misc.py
@@ -283,20 +283,3 @@ def memoize(obj):
             cache.popitem(last=False)
         return cache[key]
     return memoizer
-
-
-if __name__ == '__main__':
-    if os.name == 'nt':
-        assert get_common_path([
-                                'D:\\Python\\spyder-v21\\spyder\\widgets',
-                                'D:\\Python\\spyder\\spyder\\utils',
-                                'D:\\Python\\spyder\\spyder\\widgets',
-                                'D:\\Python\\spyder-v21\\spyder\\utils',
-                                ]) == 'D:\\Python'
-    else:
-        assert get_common_path([
-                                '/Python/spyder-v21/spyder.widgets',
-                                '/Python/spyder/spyder.utils',
-                                '/Python/spyder/spyder.widgets',
-                                '/Python/spyder-v21/spyder.utils',
-                                ]) == '/Python'
