+++ spyder/utils/tests/test_programs.py
@@ -10,7 +10,9 @@ import pytest
 
 from spyder.utils.programs import (run_python_script_in_terminal,
                                    is_python_interpreter,
-                                   is_python_interpreter_valid_name)
+                                   is_python_interpreter_valid_name,
+                                   find_program, shell_split, check_version,
+                                   is_module_installed)
 
 
 if os.name == 'nt':
@@ -68,6 +70,27 @@ def test_is_valid_interpreter_name():
     names = ['python', 'pythonw', 'python2.7', 'python3.5', 'python.exe', 'pythonw.exe']
     assert all([is_python_interpreter_valid_name(n) for n in names])
 
+def test_find_program():
+    """Test if can find the program."""
+    assert find_program('git')
+
+def test_shell_split():
+    """Test if the text can be split using shell-like sintax."""
+    assert shell_split('-q -o -a') == ['-q', '-o', '-a']
+    assert shell_split('-q "d:\\Python de xxxx\\t.txt" -o -a') == \
+           ['-q', 'd:\\Python de xxxx\\t.txt', '-o', '-a']
+
+def test_check_version():
+    """Test the compare function for versions."""
+    assert check_version('0.9.4-1', '0.9.4', '>=')
+    assert check_version('3.0.0rc1', '3.0.0', '<')
+    assert check_version('1.0', '1.0b2', '>')
+
+def test_is_module_installed():
+    """Test if a module with the proper version is installed"""
+    assert is_module_installed('qtconsole', '>=4.0')
+    assert not is_module_installed('IPython', '>=1.0;<3.0')
+    assert is_module_installed('jedi', '>=0.7.0')
 
 if __name__ == '__main__':
     pytest.main()
